+++ src/parser/tests/statement/Misc.spec.ts
@@ -265,7 +265,6 @@ describe('parser', () => {
         expect(statements[0].func.body.statements[1].value.elements[1].text).to.equal('rem: 2');
         expect(statements[0].func.body.statements[2].value.elements[0].text).to.equal('rem: 3: name: "bob"');
         expect(statements[0].func.body.statements[3].name.text).to.equal('rem');
-
     });
 
     it('handles quoted AA keys', () => {